Moved utils.py to a folder.
