+++ acme_diags/driver/polar_driver.py
@@ -115,8 +115,8 @@ def run_diag(parameter):
                 plev = parameter.plevs
                 print('Selected pressure level: {}'.format(plev))
 
-                mv1_p = utils.convert_to_pressure_levels(mv1, plev, test_data, var, season)
-                mv2_p = utils.convert_to_pressure_levels(mv2, plev, test_data, var, season)
+                mv1_p = utils.general.convert_to_pressure_levels(mv1, plev, test_data, var, season)
+                mv2_p = utils.general.convert_to_pressure_levels(mv2, plev, test_data, var, season)
 
                 # Select plev.
                 for ilev in range(len(plev)):
@@ -126,7 +126,7 @@ def run_diag(parameter):
                     for region in regions:
                         print("Selected region: {}".format(region))
 
-                        mv1_domain, mv2_domain = utils.select_region(
+                        mv1_domain, mv2_domain = utils.general.select_region(
                             region, mv1, mv2, land_frac, ocean_frac, parameter)
 
                         parameter.output_file = '-'.join(
@@ -136,7 +136,7 @@ def run_diag(parameter):
 
                         # Regrid towards the lower resolution of the two
                         # variables for calculating the difference.
-                        mv1_reg, mv2_reg = utils.regrid_to_lower_res(
+                        mv1_reg, mv2_reg = utils.general.regrid_to_lower_res(
                             mv1_domain, mv2_domain, parameter.regrid_tool, parameter.regrid_method)
 
                         # Plotting
@@ -147,7 +147,7 @@ def run_diag(parameter):
                         parameter.var_region = region
                         plot(parameter.current_set, mv2_domain,
                              mv1_domain, diff, metrics_dict, parameter)
-                        utils.save_ncfiles(
+                        utils.general.save_ncfiles(
                             parameter.current_set, mv1_domain, mv2_domain, diff, parameter)
 
 
@@ -156,7 +156,7 @@ def run_diag(parameter):
                 for region in regions:
                     print("Selected region: {}".format(region))
 
-                    mv1_domain, mv2_domain = utils.select_region(
+                    mv1_domain, mv2_domain = utils.general.select_region(
                         region, mv1, mv2, land_frac, ocean_frac, parameter)
 
                     parameter.output_file = '-'.join(
@@ -165,7 +165,7 @@ def run_diag(parameter):
 
                     # Regrid towards the lower resolution of the two
                     # variables for calculating the difference.
-                    mv1_reg, mv2_reg = utils.regrid_to_lower_res(
+                    mv1_reg, mv2_reg = utils.general.regrid_to_lower_res(
                         mv1_domain, mv2_domain, parameter.regrid_tool, parameter.regrid_method)
 
                     # Special case.
@@ -185,7 +185,7 @@ def run_diag(parameter):
 
                     plot(parameter.current_set, mv2_domain,
                          mv1_domain, diff, metrics_dict, parameter)
-                    utils.save_ncfiles(parameter.current_set,
+                    utils.general.save_ncfiles(parameter.current_set,
                                        mv1_domain, mv2_domain, diff, parameter)
 
             else:
