+++ acme_diags/plot/cartopy/cosp_histogram_plot.py
@@ -6,7 +6,7 @@ import matplotlib
 matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import matplotlib.colors as colors
-from acme_diags.driver.utils import get_output_dir, _chown
+from acme_diags.driver.utils.general import get_output_dir, _chown
 from acme_diags.plot import get_colormap
 
 plotTitle = {'fontsize': 11.5}
