+++ acme_diags/plot/cartopy/polar_plot.py
@@ -9,7 +9,7 @@ import matplotlib.pyplot as plt
 import matplotlib.colors as colors
 import cartopy.crs as ccrs
 import matplotlib.path as mpath
-from acme_diags.driver.utils import get_output_dir, _chown
+from acme_diags.driver.utils.general import get_output_dir, _chown
 from acme_diags.plot import get_colormap
 
 plotTitle = {'fontsize': 11.5}
@@ -138,7 +138,6 @@ def plot(reference, test, diff, metrics_dict, parameter):
     fig = plt.figure(figsize=parameter.figsize, dpi=parameter.dpi)
 
     # Create projection
-    print(parameter.var_region)
     if parameter.var_region.find('N') != -1:
         pole = 'N'
         proj = ccrs.NorthPolarStereo(central_longitude=0)
